Move some non-public API to `src`
