+++ client/lib/src/target_builder.dart
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019, TeamDev. All rights reserved.
+ * Copyright 2020, TeamDev. All rights reserved.
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -19,39 +19,34 @@
  */
 
 import 'package:protobuf/protobuf.dart';
-import 'package:spine_client/actor_request_factory.dart';
-import 'package:spine_client/spine/core/command.pb.dart';
+import 'package:spine_client/spine/client/filters.pb.dart';
 import 'package:spine_client/src/any_packer.dart';
-import 'package:spine_client/uuids.dart';
+import 'package:spine_client/src/known_types.dart';
 import 'package:spine_client/validate.dart';
 
-/// A factory of commands to send to the server.
-class CommandFactory {
-
-    final ActorProvider _context;
-
-    CommandFactory(this._context);
-
-    /// Creates a command with the given message.
-    Command create(GeneratedMessage message) {
-        checkValid(message);
-        var cmd = Command();
-        cmd
-            ..id = _newId()
-            ..message = pack(message)
-            ..context = _buildContext();
-        return cmd;
+/// Creates a target which matches messages with the given IDs.
+Target target(GeneratedMessage instance,
+              {Iterable<Object> ids,
+               Iterable<CompositeFilter> fieldFilters}) {
+    var target = Target();
+    target.type = _typeUrl(instance);
+    var filters = TargetFilters();
+    if (ids != null && ids.isNotEmpty) {
+        var idFilter = IdFilter();
+        idFilter.id.addAll(ids.map(packId));
+        filters.idFilter = idFilter;
     }
-
-    CommandContext _buildContext() {
-        var ctx = CommandContext();
-        ctx.actorContext = _context();
-        return ctx;
+    if (fieldFilters != null && fieldFilters.isNotEmpty) {
+        filters.filter.addAll(fieldFilters);
     }
-
-    CommandId _newId() {
-        var id = CommandId();
-        id.uuid = newUuid();
-        return id;
+    if (isDefault(filters)) {
+        target.includeAll = true;
+    } else {
+        target.filters = filters;
     }
+    return target;
+}
+
+String _typeUrl(GeneratedMessage message) {
+    return theKnownTypes.typeUrlOf(message);
 }
